update on feedInventory entity

+++ ChickTrack.Domain/DataTransferObjects/Feed/FeedLogDto.cs
@@ -7,8 +7,8 @@ namespace ChickTrack.Domain.DataTransferObjects.Feed
         public long Id { get; set; }
         public string Code { get; set; }
         public FeedBrandEnum FeedBrand { get; set; }
-        public int BagsBought { get; set; }
-        public int BagsSold { get; set; }
-        public int AvailableBags { get; set; }
+        public decimal BagsBought { get; set; }
+        public decimal BagsSold { get; set; }
+        public decimal AvailableBags { get; set; }
     }
 }
